+++ src/game/tests/move-logic.test.ts
@@ -1,5 +1,4 @@
 import { describe, expect, it } from "@jest/globals";
-import { GHQState } from "@/game/engine";
 import { movesForActivePiece } from "@/game/move-logic";
 import {
   Blue,
